<feat> spilt service logic
비지니스, 서비스 로직을 구분하여 리팩토링 진

close #40
