--- requirements.txt
@@ -1,127 +1,0 @@
-absl-py==1.0.0
-alembic==1.7.7
-anyio==3.5.0
-appdata==2.1.2
-appnope==0.1.3
-asgiref==3.5.1
-asttokens==2.0.5
-astunparse==1.6.3
-async-timeout==4.0.2
-attrs==21.4.0
-autopage==0.5.0
-backcall==0.2.0
-cachetools==5.0.0
-certifi==2021.10.8
-cffi==1.15.0
-charset-normalizer==2.0.12
-click==8.0.3
-cliff==3.10.1
-cmaes==0.8.2
-cmd2==2.4.1
-colorlog==6.6.0
-cryptography==37.0.2
-debugpy==1.6.0
-decorator==5.1.1
-Deprecated==1.2.13
-distlib==0.3.4
-entrypoints==0.4
-executing==0.8.3
-fastapi==0.76.0
-filelock==3.6.0
-flatbuffers==2.0
-gast==0.4.0
-google-auth==2.6.6
-google-auth-oauthlib==0.4.6
-google-pasta==0.2.0
-greenlet==1.1.2
-grpcio==1.46.0
-h11==0.13.0
-h5py==3.6.0
-hiredis==2.0.0
-idna==3.3
-importlib-metadata==4.11.3
-ipykernel==6.13.0
-ipython==8.3.0
-jedi==0.18.1
-joblib==1.1.0
-jupyter-client==7.3.0
-jupyter-core==4.10.0
-keras==2.7.0
-Keras-Preprocessing==1.1.2
-libclang==14.0.1
-Mako==1.2.0
-Markdown==3.3.6
-MarkupSafe==2.1.1
-matplotlib-inline==0.1.3
-nest-asyncio==1.5.5
-numpy==1.22.3
-oauthlib==3.2.0
-onnx==1.11.0
-onnxconverter-common==1.9.0
-opt-einsum==3.3.0
-optuna==2.10.0
-packaging==21.3
-pandas==1.4.2
-parso==0.8.3
-pbr==5.9.0
-pexpect==4.8.0
-pickleshare==0.7.5
-pipenv==2022.5.2
-platformdirs==2.5.2
-postgres==4.0
-prettytable==3.3.0
-prompt-toolkit==3.0.29
-protobuf==3.20.1
-psutil==5.9.0
-psycopg2-binary==2.9.3
-psycopg2-pool==1.1
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-pycparser==2.21
-pydantic==1.9.0
-Pygments==2.12.0
-PyMySQL==1.0.2
-pyparsing==3.0.8
-pyperclip==1.8.2
-python-dateutil==2.8.2
-pytz==2022.1
-PyYAML==6.0
-pyzmq==22.3.0
-redis==3.5.3
-redisai==1.2.2
-requests==2.27.1
-requests-oauthlib==1.3.1
-rsa==4.8
-scikit-learn==1.0.2
-scipy==1.8.0
-six==1.16.0
-skl2onnx==1.11.1
-sniffio==1.2.0
-SQLAlchemy==1.4.36
-stack-data==0.2.0
-starlette==0.18.0
-stevedore==3.5.0
-tensorboard==2.8.0
-tensorboard-data-server==0.6.1
-tensorboard-plugin-wit==1.8.1
-tensorflow==2.7.0
-tensorflow-estimator==2.7.0
-tensorflow-io-gcs-filesystem==0.25.0
-termcolor==1.1.0
-tf-estimator-nightly==2.8.0.dev2021122109
-threadpoolctl==3.1.0
-to-requirements.txt==1.1.8
-tornado==6.1
-tqdm==4.64.0
-traitlets==5.1.1
-typing_extensions==4.2.0
-urllib3==1.26.9
-uvicorn==0.17.6
-virtualenv==20.14.1
-virtualenv-clone==0.5.7
-wcwidth==0.2.5
-Werkzeug==2.1.2
-wrapt==1.14.1
-zipp==3.8.0
